Change the fSecs field of plUnifiedTime from UInt32 to time_t.
This fixes various date formatting problems when building on a system where time_t is 64-bit (e.g. Visual Studio 2010), and, as a bonus, extends the range past 2038 on such systems.

The wire protocol is left at 32-bit for now, we might change that to 64 when other reasons to break compatibility have accumulated.