Added DEPR_STR macro to help plStringification.
You pass this as the first argument to a function that should be
deprecated in favour of a plString version. By default, the macro
expands to nothing; but you can set NO_OLD_STRINGS in CMake which will
cause it to add a void* parameter, causing all calls to that function to
break.