Fullscreen clients no longer hide message boxes
The old solution may have worked at one point in time, but it was kind of
suspect in that it would (theoretically) hide the game window forever. The
new solution will only hide the game window while the message box is
alive. As a bonus, the part where we hide the full screen window now
works.