Fix Zandi Welcome
Due to skoader's physics changes, Zandi triggers the avatar detection
region when he spawns. This caused him to welcome himself to the cleft but
left the actual avatar without a chance to hear the welcome message. This
changeset ensures that the triggering avatar will play the speech logic on
collisions.