Another decimal place in the profiling display
And Hoikas said, "I cannot tell a difference in the skinning times on my
not-crappy computer with only a tenth of a millisecond resolution." So we
now have hundredth of a millisecond in the profiling display. These are
the kinds of problems you have when developing with a machine that doesn't
suck. It would be nice if players didn't try to run the game on machines
from before the exctinction event that ended the Permian era.