File download retries...
Ugh. So, some servers really suck and claim they don't have some file--BUT
THEY REALLY DO! I will never understand this, so, as a hack, we will retry
the same file up to three times. To prevent this from being a
spin-error-boom loop a phailure, we'll retry files at the very end of the
patch proc.

See also:
http://www.mystonline.com/forums/viewtopic.php?p=408060#p408060