Only dirty changed states...
So the big idea for dirtying a state is that we need to send it to the
server, because, well, we changed it. So, let's not keep everything dirty.
We can figure out if it's changed from the default regardless of the dirty
flag.

This change specifically fixes problems with dirtied-but-not-modified
variables overriding global states (AllAgeGlobalSDL) just because they
happened to be read in at a later time than the global state. Whew.