Revert "Fix dupe callbacks in VaultAddChildNode"
This reverts commit b393947cd596ce461b887b6f02fc54d31ec94c63.

The commit was seemingly harmless, but it created or otherwise uncovered
stack corruption deep inside the vault voodoo. While I would like to know
what was up, I'm tired of dealing with it. The crash addressed here was
fixed in a subsequent moul-scripts commit, so let's just toss this one.