Check if CMake is accessible before trying to invoke it.
The default option in CMake's installer is to not install itself into the path.  Our directions on http://h-uru.github.io/Plasma/ remind the user to select the correct option, but they may overlook this, and be faced with an error running this script.  Adding a specific message here during this case will help guide them to self-fixing the problem.