Convert vault tree printing to plString
After hacking the old to just barely use plString, I received some UTF-8
asserts. So, that seemed like a good excuse to convert the whole mess to
plString. Note that the output is now saved in VaultClient.log. This
should help with getting some of the chaos out of OLD_ASYNC_LOG.log