Always display the bottom-bar KI alert for GZ Markers in range.
The change on the MiniKI display from a white to red marker is easy to miss, so
in order to see the easily-visible KI alert in the bottom bar the player has to
constantly close the KI after each marker capture.

This minor fix always displays the alert even if the MiniKI is open; it reduces
tedium and makes it easier to do marker missions with less useless clicking and
even while chatting.