Update the chat area intelligently
Again, refreshing the KI dialog is expensive. There are actually many
implicit updates to the MLEdit control when doing very common operations.
Now, we batch all our updates in and tell the control when we're done so
that it's only redrawn once. I also took the opportunity to remove a call
to insertStringW, whose code looks quite insane IMO.