Fix CGZM times not updating correctly
This fixes an issue in which the calibration missing display in the Great
Zero age would not update when a new score is recorded. As a bonus, we fix
a potential issue that would cause a mission to be completed successfully
if the user reopens the GUI when a mission is active and no markers have
been captured.