Fix PhysX exception swallowing
Since the VC++2015 update, it has been extremely difficult to attach the
debugger to a crash[ed|ing] plClient due to PhysX's buggy unhandled
exception filter. All that it appeared to do anyway was tell the
application to close, so unhooking it is for the best. We can now get
minidumps more reliably and attach the debugger.