Fix issues with repair mode patching
As it turns out, you cannot open 2,000 files simulatenously on Windows.
The end result of attempting to do so is lots of data downloaded from
the server but saved to an invalid fd. This results in bugs where the
player has to patch, patch, patch, and repatch :(. Also, sprinkle in
some HS_OVERRIDE for fun.