Implement patcher file deletion
Deletion is handled with a new flag added to the FileSrv manifest 0x20.
The content aside from the client entry name is irrelevant, the file is
never requested, only deleted.

Because the deletion notification comes from the patcher thread, I had
to move the logic for managing PRP removals into the plClient message
receive to prevent race conditions. This is an altogether better
situation because casting the hsResMgr outside of plClient is something
I consider evil.