Decouple patcher quit from WM_QUIT
It has been observed that sometimes zombie patchers have been lingering
around as of at least Gehn 25. It appears that we do not always receive
WM_QUIT in the program's main loop, but it is always pumped to the
DialogProc. Windows is probably translating another message somewhere
into a WM_QUIT for us and posting it directly from the DialogProcedure.

Thus, I have decoupled the patcher's main loop quit from receiving a
WM_QUIT message to waiting for the networking core to shut down. This is
done when WM_QUIT is received in the DialogProc, which is highly
reliable. In my testing, this seems to fix the zombie patchers...