Support libpng with no iTXt chunk support
libpng's support for the `iTXt` chunk ('international textual data') is
optional, and off-by-default before libpng 1.3, which causes a compile
error. This writes an iTXt chunk only if supported, and writes a tEXt
chunk otherwise. (A tEXt chunk is identical but specified to contain
Latin-1 text instead of UTF-8, and has no language tag.)