Add more robust aspect ratio handling
Ultrawide displays are generally advertised as having a 21:9 aspect
ratio, even when they're a bit off from that mathematically speaking.
So, we now allow the text for the aspect ratio to be set separate from
the numerical values. Also changed the pixel fudge calculations to a
range of aspect ratios.