Add cmake config options for opus and speex.
This adds yet another cmake config file. I decided to not use the other
one because it is for detecting compiler features. A good future project
would be to pull out a lot of the project level definitions out and move
them into the limited-scope hsConfig.h. This should minimize the amount
of recompilation needed in, for example, changing from internal to
external releases.