Third pass of plPythonFileMod cleanups.
This changes the way script methods are called. Instead of the verbose
calling we did previously, we now have a variadic template powered
calling function for PFM methods. I had to change fPyFunctionInstances
to actually hold a pointer to the methods instead of a pointer to the
PFM instance... Who thought it was a good idea to have a 30+ member
array that points to either fInstance or null???