Fix microphone volume level control.
Previously, the "My Voice" control in the options GUI did nothing. The
old plWinMicLevel code was originally for the WinMM library from 1991
(!!!). This rewrites all that junk to use the Windows Core Audio library
upon which WinMM sits as of Windows Vista. When using OpenAL Soft, the
primary backend is WASAPI, a Core Audio component. Further, when
adjusting the level in Uru, you can observe the level changing in the
Windows mixer as well.