Fix ptVaultNode's richcompare function.
The Cyan standard function did not properly handle comparing vault nodes
with incompatible types, which is expected to return False. Instead,
they tried to return an exception. Unfortunately, they did this
incorrectly by returning a value of -1, which was interpreted to be a
result of True.

This resulted in interesting things like `node == "Ahnonay"` evaluating
to True.