Bump to Python 3.8.
This is required due to reports from multiple testers that the client
exits on loading StartUp. The issue is that Python 3 loads modules
during its initialization--before we have installed our PEP 451 import
machinery for python.pak. The previous code worked for us because we
have Python 3 installed on our systems, which is not a reasonable
assumption to have for our players. Python 3.8 adds APIs that allow us
to initialize just the "core" and use the API, then initialize the
interpreter itself once we have added the import machinery.