prepare_env script improvements
Set PowerShell 5.0 as a minimum required version to be able to use
modern commands for downloading and unzipping the devlibs.

Also don't assume Windows path separators. This allows the script to run
(to the point of failing to detect CMake) on PowerShell for Linux.

Also ensure CMake generates projects that target Win32 instead of Win64.