Return uname info for DisplaySystemVersion on *nix
This gives us the kernel version and machine architecture.

Note that on macOS, this will return "Darwin" and the Darwin kernel
version, rather than the (more useful) macOS product version. I looked
into getting that info as a second item in the returned vector, but the
easy way to do it involves private Apple APIs and the less-easy way to
do it involves using CoreFoundation to parse plist files and extract
values from a dictionary. Not impossible, but not exactly fun either.