Rewrite the character controller as a dynamic actor.
With the PhysX CCT, it was too easy to get into situations where the
avatar could ride on top of a kickable to access out-of-bounds areas.
Additionally, the hit force applied in those situations was at best
comical and at worst frustrating due to the physics based puzzles.

This rewrites the character controller to be a true dynamic rigid body
in the simulation controlled by setting its linear velocity each
simulation frame. The result is greatly simplified code and better
interaction with the world.