Fix fallout from upstream libexpat library rename.
As of libexpat 2.2.8, the library was renamed from libexpat to expat. In
2.2.9, this was supposedly reverted back to libexpat; however, in practice,
I find that vcpkg builds libexpatMD. Therefore, we now use libexpat's
config. However, this is not present in the most recent devlibs bundle
with expat 2.2.7, so we fall back on the out of date cmake find module
in that case.