Fix issues with ladders throwing avatars into space.
Apparently, if collisions occur while we are forcibly setting the global
pose of an actor, this causes some kind of internal force accumulator
to, well, accumulate. Once we resume normal operation, this accumulator
causes the avatar to be flung into space. Simply setting the linear
velocity to zero does not fix the issue.

With that in mind, we now expose another flags field on the simulation
shader and selectively suppress collisions with non-triggers when the
avatar brain thinks there should be no physics.