Fix a crash when writing error messages to the log.
This crash was reported in IRC by a user running French Windows. The
client would crash almost immediately upon the login dialog becoming
visible. Upon investigation, string_theory was throwing an exception
from formatting an `std::string` coming from
`std::error_category::message`. As best I can tell, the MSVCRT
implementation of this facility is returning the error message encoded
in the current codepage -- or windows-1252 in the case of French
Windows. This works OK on English Windows, but we need to use proper
unicode for other variants.