Fix non-synchonized link out animation.
This fixes the issue by which the linking responder in a non-bookshelf
book is not properly synchronized to remote players. Preciously, On the
local machine, the player sees him or herself placing his/her hand on the
page and linking out. On the remote machine, an avatar walks up to the
linking book and links away without touching the page.