Refactor plShaderAssembler to use D3DCompiler.
D3DX is gone, so we have to use the standard D3DCompiler library. Sadly,
the D3DCompiler only publicly offers an HLSL Compiler and not an
assembler. However, after disassembling d3dx, I found that D3DX just
dispatches assembler calls into an undocumented export in D3DCompile. We
now use this to assemble the shaders from the assembly source.

If someone wants to add support for HLSL shaders, it should probably be
done in a separate program.