Update resource.dat requirements to a better cairosvg.
The latest version of cairosvg on pip seems to have an undocumented
dependency that either cairo or GTK is installed and available on the
PATH. Therefore, we roll back to the old cairosvg 1.x on all Python
versions. This version of cairosvg using pycairo as a backend and works
correctly, without hidden surprises.