Prevent automatic unpacking of hsRef via operator overload.
Unpacking must be explicitly done via Get() now, which prevents accidental
misuse in packing and unpacking of hsRef-wrapped objects.

This also makes greater use of hsWeakRef to clearly indicate where Vault
objects are used as a weak reference (instead of a raw pointer).  For now,
the line is drawn at the NetCli* level, but this could be extended further
in the future.