Properly generate resource.dat as a plClient dependency.
This does a lot of things. On a high level, it changes the resource.dat
file generation such that it properly gets marked as a GENERATED
dependency of plClient. Further, we correctly chain the vector graphics
and scripts as dependencies of resource.dat such that resource.dat will
only be regenerated by modifying these files. #745 unintentionally broke
this by removing the default definition of
`CMAKE_RUNTIME_OUTPUT_DIRECTORY` -- CI didn't detect it because it it
defines that variable in the cache.

More sneakily, we now see the beginnings of more automated dependency
management in that our CMake can now install the python packages for
building resource.dat. For now, it is OFF by default - it seems rude to
unilaterally install packages in a system Python install. Future work to
enable a submodule based vcpkg toolchain would use this facility
however.