First major pass over target_link_libraries.
This fixes the majority of #383. In order to do so, I began to reorder
some of the include directives as specified in #788 to facillitate
easier inspection of the dependencies. The work is not complete at all.
Further, I broke several dependencies where trivially possible by
changing uses of derived classes to base classes and moving virtual
function definitions into source files from headers.

The diff is unpleasant, I know :(