Don't use the QFileDialog native dialog backend in the Qt tools.
Qt apparently has a long-standing bug that causes the native file dialog
on Windows to hang indefinitely if the application initializes COM
(CoInitialize[Ex]) before the dialog is shown.  Since Plasma does this
for us, all file dialogs are broken with the native backend on Windows.
The non-native dialog may not look as well integrated, but it works.