Fix SIMD usage on GCC/clang.
If the -m(instruction set) flag is not set when compiling a source file,
then intrinsics from that instruction set are not available. This means
that when parsing a source file that uses, eg, SSE3, then the file will
fail to compile. Unfortunately, we can't just turn on every instruction
set we want to use because that will allow the compiler to emit the
given instruction set anywhere. This will result in fun-to-debug
illegal instruction crashes for our end users.

Anyway, if you have a file that needs to compile with gcc, be sure to
jail any SIMD code to its own file and add it to the target using
`plasma_target_simd_sources(TARGET [instruction set] [files...])`.