Fix a crash introduced by c19ae73a:
The pool vector only calls the createItem code when adding a new element
to the pool, but plDXPipeline will explicitly null out items in the pool.
This change adds a reset method which also clears the pool (note: per the
C++17 standard, clear() will not cause a reallocation of the underlying
storage, so this is not a performance regression).