Implement pfConfirmationMgr.
pfConfirmationMgr is designed to replace the ad-hoc Yes/No dialog
implementation in xKI. Currently, the KI handles throwing up its own
dialogs and can respond to pfKIMsgs sent with a kYesNo type. The result
is the handling of confirmation dialogs is:

- Dependent on having a functioning KI python.
- Spread over many hundreds of lines of python.
- Requires having a plKey to receive a response.

These limitations make doing things that require confirmation annoying.
This system will allow you to dispatch a confirmation dialog from
anywhere in the engine that links against plMessage and receive the
results either in a (potentially stateful) lambda or via a legacy
notification message.