Censor fixes for URLs.
We cannot censor the contents of KI mail in the vault. This would cause
URLs containing naughty words to be irrevocably broken, linking to
somewhere other than the original link. As a side effect, we now lock
the editing of KI mail to its creator (or sender) to prevent naughty
words escaping the censor.