Fix wild camera LOS issues.
The camera LOS had two issues that could cause the camera to go to
either odd locations or swing wildly around 180 degrees near obstacles.
For the former, we were potentially leaking our internal state in LOS
miss responses, causing interesting camera goals. For the latter, we
were erronously reporting hits at a distance of zero. These hits mean
"you're on top of it." In reality, we don't care about that. Besides,
what is a ray of length zero?

Anyway, this refactors the whole mess to be less fiddly and use only one
raycast instead of potentially two. I opted to not use the cache because
of some poorly phrased commentary in the API documentation indicating
the cache is basically never used.