Fix clothing UVs not being updated by morphs.
Generally skinned objects will not change their UVs - except in the case
of avatar clothing. When the morph slider is adjusted in the ACA, new
UVs may or may not be committed to the `plGBufferGroup`. If we simply
skip over them in the skinning loop, then the newly morphed UVs will not
be sent to DirectX.