Convert all the `MAX_LOCALIZED[1-2]` macros to proxies.
These proxies are a little fiddly, but I can't think of a better way to
do it since the old overloads are now `MAX_SEALED` and proxy to the new
overload. While we could undef `MAX_SEALED`, that assumes AD will never
change `MAX_SEALED` to an outright `final`, so we'd be back to this same
problem again.