Allow the max plugin to auto-generate PlasmaMAX2.ini
Previously, if you launched Max without a PlasmaMAX2.ini file, you were
SOL. The functionality to specify a client directory was already
present, so this simply enables it. As a bonus, we now use the folder
picker dialog from Vista (and above) instead of the old Windows 9x
dialog, and windows.h becomes much less viral.