Fix crash on 3DS Max exit
The Export dialog would try to write some data to the PlasmaMax2.ini
config file when it was destructed, however the core interface to fetch
that config file seems to already be destructed at that point. It's
possible the order of operations or behaviour of dialog destruction
changed in newer Max versions.

However, it seems to make the most sense to save those values when the
Export dialog is closed, rather than when it is destructed.