Fix custom avatar footstep sounds
The age and page names for footstep sounds are stored in the
plArmatureMod data, but the effects were getting initialized before
those values were read, so they were always using the defaults of
"GlobalAvatars" and "Audio". This results in the custom footstep sounds
for Quabs not taking effect in-game.

However, this will break footstep sounds for other custom avatars who
*do* want to use the global sounds, because the age name is exported as
"CustomAvatars" instead of "GlobalAvatars" in their plLODAvatarMod data.

This commit include an update to PlasmaMax so that if the avatar
footstep page is set to "Audio" it will export with a hardcoded
"GlobalAvatars" age, otherwise it will use the current age name.

The custom avatar data can easily be fixed in the compiled data files in
the moul-assets repo.