Use `VCPKG_DISABLE_COMPILER_TRACKING` to improve caching.
With the most recent update of vcpkg-tool, vcpkg will now allow us to
skip including the hash of cl.exe with the binary cache. This means that
minor compiler revisions eg 16.10.1 -> 16.10.2 will not invalidate the
cache. So, downstream users will more likely be able to download the
libraries from CI.

This decision may need to be revisited when Microsoft breaks ABI
compatbility in vNext. However, per the STL repo, this has been delayed
indefinitely. (NOTE: Visual Studio 2022 is not vNext).