Add avatar controller contact suppression logic.
This suppresses contacts with characters who seem to be stuck falling in
place, which greatly improves the chance that the simulation will allow
them to fall to the ground. Also, remove the friction from any collision
that looks like a wall, which should reduce the liklihood of barely
side-swiping a wall completely tanking your forward velocity.