Fix a variety of friction problems.
So, apparently, the getting stuck problems were all because our force of
static friction was effectly equal and opposite the applied force of the
character's movement in some situations. We really don't want that when
running alongside walls, jumping onto objects, and sliding down cliffs.
But we do want friction to help us not run out of the world. Tough
balance!