Correct ShowSDL command to match Set commands.
The set commands operate on the Python AgeSDLHook, but the ShowSDL
command was showing the age vault SDL values. On DirtSand, the server
merges these together, however on Cyan and MOSS servers, these values
are distinct. Therefore, we separate them into their own commands. The
Set commands operate on the Python SDL.