Improve the "step-up" helper.
Previously, there were noticable hitches when the step-up feature
engaged, especially on the Kadish X2 Vault stairs. This is because the
step-up functionality was engaging in a way that would only take effect
on the next render frame. This changes it to take effect on the current
render frame, resulting in a smoother result.