Improve text resolution of BaronCityOffice Marie note
This requires a corresponding change in
BaronCityOffice_District_BCOWrinkledPaper.prp to double the size of the
dynamic text map.

The original code sets the font size to 10, but the game actually used
size 12 instead (the lowest available size of the Michelle font), so
this actually doubles the font size exactly (from 12 to 24).