Correct not-on-ground fricition.
If we're NOT on the ground, we want to enable infinite friction. Imagine
that you are on a crappy system and running across the threshold of a
staircase going down. The huge frame delta will result in your going
"airborne" - without the friction force, you will skid down the stairs
like it's made of ice!