Add RelWithPDB configuration for MSVC.
The RelWithDebInfo configuration is great for trivially getting a PDB
file, but it doesn't do the full level of optimizations that Release
does. So, this adds a new build type that sets the optimization levels
up to Release but ensures a PDB file is generated. This configuration
should only be used for actual releases. For general development level
hacking, RelWithDebInfo is more appropriate.