Fix #985.
The problem here is that when the mouse moves quickly, the movement
delta is so large that the mouse enters the "left", "right", or "down"
arrow regions briefly before the re-center logic is able to reposition
the mouse. This triggers the input code to change the cursor to one of
those appropriate images. The problem happened in that both the mouse
cursor update routine and the cursor creation routine had logic problems
that would show the cursor when it changes. The latter would even
override forced cursor hides done by `plMouseDevice::HideCursor(true)`.

I elected to NOT make panning the camera a forced hide (see
`plCameraBrain1::MsgReceive()`'s `S_SET_FREELOOK` handler) because that
seems like more of a gameplay oriented feature. It would be a shame if
the cursor got stuck force hidden because we accidentally missed a mouse
release, mouse-escaping the window, or something like that.