Fix for z-fighting with physics bodies
This should fix things like the fire marbles in Neighborhood. Those
objects contain physics bodies that render as transparent, but were
still contributing to the Z buffer and sometimes blocking other
textures. This borrows from the DX renderer, where transparent bodies
are not rendered. That also keeps them from contributing to the
Z-buffer.