Fix for alpha testing in fragment shader
This should solve z-fighting in transparent textures. I haven't compared
against DX so it's a bit of a guess, but fragments that are completely
transparent should probably be discarded. This was causing z-fighting to
occur with transparent materials that were overlapped on each other.
Commonly seen in the "HeekD'nji" material in "Neighborhood."