Opacity based culling fixes
The GetOpacity based culling needed some more conditions, it was doing
things like not rendering the fountain in Eder Kemo. That should be
resolved now, and the fire marbles are still fixed.

The alpha threshold change caused unresolvable issues in Eder Kemo.
Undoing for now. I'll have to figure out a better way to fix that
hologram.