Don't delete strings owned by Max.
I'm not sure how accurate the comment given here is. Previously, before
the code was fixed to handle UNICODE, we were indeed making a copy of
the string and passing it off to Max. However, as of b6966c9, we're
handing Max a temporary copy of the string. That clearly works now, so
the original author's understanding of how Max stored these strings was
either flawed, or Max used to be tolerant of us deleting its strings out
from under it. Either way, deleting a string owned by Max is bad in
2022, so debounce that crap.