Add margins to KI chat display to fix shadow clipping.
When chat lines did not have a leading space character, you could
sometimes see that shadows were being clipped due to drawing 2 pixels
out of bounds. While I generally think managing the shadow unclipping
would be better in the text rendering code, Plasma's text rendering is
not that smart and works on the idea that text is rendered exactly where
you specify it. So, we're largely obligated to handle this kind of text
positioning logic in higher level code.