Remove foreign swear words from the censor.
Per Zaroth, "aujourd'hui" is being censored in the French Q1 2022 Kirel
speech subtitles. "hui" is "today" in French but is vulgar slang for
penis in Russian. Given the English-centric nature of the list and the
possibility of false-friends, it seems best to simply remove foreign
words from the censor list. Future work would be to add proper censored
words for each language.