bug fixes encountered during testing on Minkata
Rearranged MaleSuit and FemaleSuit to be in the same order as clothing gathered with getAvatarClothingList()
solo variable was returning false when player was alone. We wanted it to be true when player is alone.
Added DefaultColor to when the suit is first worn, so that when its awarded the players hair color and clothing colors are not all set to white.