Adjustments made after testing on Moula Staging
grsnNexusBookMachine.py
Changed way nexus return book is deployed. Now when game master triggers, it send a onNotify to all players, instead of trying to sync the animation from the game master.

grsnWallPython.py
check added to ignore accessories as they are not removed when Maintainer suit is worn. No need to try to recolor them as they were not modified.