Adjust controller size for Teledahn.
This fixes a problem in which the camera region for the Teledahn sewer
ladder did not fire correctly. This capsule size adjustment seems to be
a leftover from when we were making a custom capsule and is no longer
appropriate. The result was that the character capsule was too thin to
fire the camera region.