Fix problem with copying not working reliably.
If you already have data in a different format in the clipboard, such as
an image, copying text into the clipboard without clearing it can result
in interesting results when attempting to paste. If the application
you're pasting into accepts both images and text (eg discord), in my
testing, the image is pasted instead of the newly copied text, leading
to confusing UX with reports of the copy button not working. Therefore,
we explicitly clear the clipboard content but only if there is an active
window on the calling thread (if there isn't one, bad things happen per
MSDN).