Remove incorrectly "fixed" obfuscation check in ReadSafeWString[Long]
The check was originally "if (name[0]* 0x80)", which is effectively
always true (because the string should never contain zero characters),
so in practice the string was always un-obfuscated. This was fine,
because WriteSafeWString always obfuscates the string.

14991dc75040af7d46e0ad98a31a3c3bee5880f7 "fixed" the check to
"if (name[0] & 0x80)", which can become false if the first character in
the (obfuscated) string is a non-ASCII character in e. g. the Latin-1
range. As a result, such strings were mis-detected as un-obfuscated.
This broke e. g. KI chat messages where the first character is
non-ASCII.

Because the original check was broken and nothing writes un-obfuscated
SafeWStrings, we can simply remove the check and unconditionally
de-obfuscate the string.