Fix OpenSSL 3 deprecation warnings from plBigNum
The "fast" BigNum prime check function is deprecated in OpenSSL 3
because it apparently doesn't do a sufficient job of testing prime-ness.
This newer method does it properly, which probably means it's slower,
but probably not to a degree that it noticeably impacts performance (if
this method is called at all, which it doesn't seem to be?).

Also, the `Div` method was complaining about a ternary that tested
whether `this` was null, which isn't possible in well-defined C++ code.
I tried to look at where this was being used to see how that could
happen and it turns out the function wasn't being used at all, so I just
removed it entirely.